requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
gspread
google-auth
openpyxl
